+++ src/libreader/intern_pool/atom.rs
@@ -1,0 +1,10 @@
+// Copyright (c) 2017, Sabre developers
+//
+// Licensed under the Apache License, Version 2.0 (see LICENSE.Apache in the
+// root directory) or MIT license (see LICENSE.MIT in the root directory),
+// at your option. This file may be copied, distributed, and modified only
+// in accordance with the terms specified by the chosen license.
+
+/// An interned string representation.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Ord, PartialOrd)]
+pub struct Atom(pub(super) u32);
